Up Date
This has some changed to allow for the removal of modified library reliance. This allows for the upload of just this code and the deflate non modified library's to work. Note that sOrbit and lOrbit pins may need to be changed if the hardware for the two sensors are reversed, simple fix being swapping A1 and A2.